Chore : CommentSlot -> CommentSubit 네이밍 변경,& 타입 분리
