+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/types/member-type.tsx
@@ -1,0 +1,10 @@
+type Role = 'ADMIN' | 'MEMBER';
+
+export type Member = {
+  role: Role;
+  userImage: string;
+  userEmail: string;
+  userName: string;
+  groupId: number;
+  userId: number;
+};
