+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/utils/task-handlers.ts
@@ -1,5 +1,5 @@
 import axiosClient from '@/lib/axiosClient';
-import { Task } from '../types/task-list-page-type';
+import { Task } from '../types/task-type';
 import useModalContext from '@/components/common/modal/core/useModalContext';
 
 export function useTaskHandlers(task?: Task) {
